Move static methods from utility to enums
