--- src/main/java/com/example/auctionapp/Util/EnumUtility.java
@@ -1,23 +1,0 @@
-package com.example.auctionapp.Util;
-
-import com.example.auctionapp.enumeration.ColorEnum;
-import com.example.auctionapp.enumeration.SizeEnum;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class EnumUtility {
-
-    public static List<String> getColors() {
-        return Arrays.stream(ColorEnum.values()).map(name -> {
-            return name.label;
-        }).collect(Collectors.toList());
-    }
-
-    public static List<String> getSizes() {
-        return Arrays.stream(SizeEnum.values()).map(name -> {
-            return name.label;
-        }).collect(Collectors.toList());
-    }
-}
