+++ src/main/java/com/example/auctionapp/enumeration/ColorEnum.java
@@ -2,6 +2,10 @@ package com.example.auctionapp.enumeration;
 
 import com.example.auctionapp.exception.BadRequestException;
 
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
 public enum ColorEnum {
     BLACK ("Black"),
     WHITE ("White"),
@@ -23,10 +27,20 @@ public enum ColorEnum {
 
     public static ColorEnum fromValue(String color) {
         for (ColorEnum value : ColorEnum.values()) {
-            if(value.label.equals(color)) {
+            if(value.getLabel().equals(color)) {
                 return value;
             }
         }
         throw new BadRequestException("Color does not exist");
+    }
+
+    public static List<String> getColors() {
+        return Arrays.stream(ColorEnum.values()).map(name -> {
+            return name.getLabel();
+        }).collect(Collectors.toList());
+    }
+
+    public String getLabel() {
+        return label;
     }
 }
