+++ src/main/java/com/example/auctionapp/enumeration/SizeEnum.java
@@ -2,6 +2,10 @@ package com.example.auctionapp.enumeration;
 
 import com.example.auctionapp.exception.BadRequestException;
 
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
 public enum SizeEnum {
     EXTRA_SMALL ("Extra Small"),
     SMALL ("Small"),
@@ -18,10 +22,20 @@ public enum SizeEnum {
     public static SizeEnum fromValue(String size) {
 
         for (SizeEnum value : SizeEnum.values()) {
-            if(value.label.equals(size)) {
+            if(value.getLabel().equals(size)) {
                 return value;
             }
         }
         throw new BadRequestException("Size does not exist");
+    }
+
+    public static List<String> getSizes() {
+        return Arrays.stream(SizeEnum.values()).map(name -> {
+            return name.getLabel();
+        }).collect(Collectors.toList());
+    }
+
+    public String getLabel() {
+        return label;
     }
 }
